requirements.txt
pymongo==3.6.1
pygithub
pylint